Make status an alias for whoami and refactor utils
