+++ src/lib/profile.ts
@@ -1,0 +1,5 @@
+const getCurrentProfile = (): string => {
+  return process.env.AWS_PROFILE || '';
+};
+
+export { getCurrentProfile };
