+++ src/lib/time.ts
@@ -1,0 +1,5 @@
+import { promisify } from 'util';
+
+const timeout = promisify(setTimeout);
+
+export { timeout };
