+++ test/command/whoami.test.ts
@@ -3,7 +3,7 @@ import { assumedRole } from '../../src/command/whoami';
 describe('whoami', () => {
   test('assumedRole', () => {
     expect(
-      assumedRole('arn:aws:sts::736289482636:assumed-role/foo-bar/aws-sdk-js-1634417672672')
+      assumedRole('arn:aws:sts::111111111111:assumed-role/foo-bar/aws-sdk-js-1111111111111')
     ).toEqual('foo-bar');
   });
 });
